+++ packages/server/src/services/sessionExecution.js
@@ -9,18 +9,8 @@ import { VCRAgentAdapter } from '../agents/vcr/VCRAgentAdapter.js';
 import { isE2ESpawnCaptureEnabled } from './e2eSpawnCapture.js';
 export { buildQueryParams } from './queryParamBuilder.js';
 import { buildQueryParams } from './queryParamBuilder.js';
-import {
-  buildPromptWithAttachments,
-} from './sessionPrompts.js';
-import {
-  activeSessions,
-  activeConversationIds,
-  handleStreamEvent,
-  handleTurnCompletion,
-  handleSessionError,
-  cleanupSessionState,
-  broadcastSessionStatus,
-} from './streamEventHandler.js';
+import { buildPromptWithAttachments } from './sessionPrompts.js';
+import { activeSessions, activeConversationIds, handleStreamEvent, handleTurnCompletion, handleSessionError, cleanupSessionState, broadcastSessionStatus } from './streamEventHandler.js';
 import { shouldRescheduleOnError, _checkProactiveReschedule } from './sessionErrors.js';
 import { schedulerService } from './schedulerService.js';
 import { buildConversationContextForModelSwitch, buildConversationContextForContinuation } from './conversationContext.js';
